fix(ENG-MM-INPUT-PIPELINE): store the Qwen3-VL, Gemma-4 and GGUF vision towers in the checkpoint's own bf16 (#1359)
Every vision tower in this tree was read from an all-BF16 checkpoint, widened to
host `std::vector<float>`, and then narrowed straight back to bf16 before its
first GEMM. The f32 was a carrier: it bought nothing and cost exactly 2x the
checkpoint's bytes. On Qwen3-VL-4B that is 830,695,424 B on disk held as
1,661,390,848 B resident.

Under `.agents/specs/vision-tower-dtype-polarity.md`, whose §6 thresholds were
declared before any number existed. Upstream has no ViT dtype of its own:
`Qwen3_VisionTransformer.dtype` IS `patch_embed.proj.weight.dtype`
(`qwen3_vl.py:633-634`) under `set_default_torch_dtype(model_config.dtype)`
(`base_loader.py:53`), at parity pin 5559679229bc.

This lands the Qwen3-VL half. `Qwen3VLVisionWeights` — and therefore
Qwen3-VL-4B, the Qwen3.5/3.6-27B dense path and the Qwen3.6-35B MoE path, which
share `LoadQwen3VLVisionWeights` — plus `Gemma4VisionWeights` and both GGUF
tower loaders now store raw bf16 bits. `LoadVisionF32` is `LoadVisionBf16`;
`MakeDevBf16` is a straight `Copy` of the stored bytes instead of a per-weight
scratch allocation and a per-weight `F32ToBF16` pass, which is why the change
can only remove work. The GGUF loaders keep `DequantGgufRowToF32` and narrow
once with the same `vt::F32ToBF16` at load time rather than at upload time.

NO OUTPUT BIT MOVES, and that is mechanical rather than hoped for. `BF16ToF32`
is a 16-bit shift, so `F32ToBF16`'s round-to-nearest-even addend cannot carry
back into bit 16. `test_vision_tower_dtype` proves it over all 65,536 patterns,
and names the 126 signalling NaNs that the explicit quieting branch does move
rather than filtering them out.

That same argument makes the spec's §6.4 mutation the wrong instrument, which is
worth saying because it changes what has to be gated. Truncating the narrow was
supposed to red an exact arm; it cannot, because on a value that originated bf16
truncation and round-to-nearest-even agree by construction. The GGUF towers ARE
exposed — they narrow a genuine f32 dequant — and neither `test_clip_mmproj_gguf`
nor `test_minimax_h3` could see it either, since both sides of their value
checks call `vt::F32ToBF16` and so pin consistency rather than the rule. The
fourth case in the new suite pins the rule itself on four f32 patterns whose low
16 bits are non-zero, stated independently of the function; replacing the
rounding with `u >> 16` reds it at 2 of 4 assertions.

Two weights deliberately stay host f32 and now say why beside themselves:
`Qwen3VLVisionWeights::pos_embed_w` and
`Gemma4VisionWeights::position_embedding_table`. Both reach host arithmetic
before anything narrows them — a bilinear gather and sum, and an x+y row sum —
so narrowing the store would move the result. Both are wider than upstream,
which casts to the model dtype at `qwen3_vl.py:335`, and reconciling them is
listed under the spec's `## Owed`.

Muse Glimmer does NOT ride, and #1359 stays open for it. Its tower is
storage-only on the production path and is not storage-only on its gate path:
`compute_dtype = kF32` computes on the stored values, and that arm's weights are
a synthetic f32 LCG that `scripts/mm/muse_glimmer_vision_ref.py:52-61` builds as
`torch.float32` and never rounds. Measured on a scratch tree rather than
predicted — the five f32 stages move from rel_l2 1.0-3.0e-07 to 2.164e-03,
2.193e-03, 2.220e-03, 2.892e-03 and 3.462e-03 against a 1e-6 bound, while the
bf16 arm reads `rel_l2=5.951e-03 max_abs=3.675e-02`, byte-for-byte what it reads
today. The widening is removable; the golden has to move with it, which is a
reference change that must not ride in the change it gates. #2166 owns it.

`scripts/mm/tower_skip_rss.sh`'s `TOWER_RESIDENT_BYTES` is therefore per-kind
instead of a blanket `* 2`, mirrored by `WIDEN` in its report suite. Keeping the
surviving Muse Glimmer `* 2` visible is what stops that half being forgotten.
The published Qwen3-VL tower-skip saving falls from 1,655,791,616 B (1.542 GiB)
to about 830,695,424 B (0.7736 GiB) as a result, and every surface that carries
that figure now says the fall is CORRECT rather than a regression: the flag now
frees the tower the checkpoint ships instead of the tower plus our widening. The
2026-08-24 measurement itself is untouched, because it is the record of a run
that happened at `41ab550b9`.

Nothing is measured here. §6.1's peak-RSS gate needs a leased host and the
staged checkpoint, and it stays PENDING with its pre-declared thresholds intact.

FOLLOWING_AGENTS_PROTOCOL

Following-Agents-Protocol: true
AI-Assisted: true

+++ include/vllm/model_executor/models/gemma4_vision.h
@@ -74,22 +74,32 @@ struct Clip {
   float out_min = -3.4e38f, out_max = 3.4e38f;
 };
 
-// All weights host-side row-major f32 (as torch stores: Linear weight [out,in]).
+// All weights host-side row-major RAW BF16 BITS (`uint16_t`, as torch stores:
+// a Linear weight is [out,in]). bf16 because `Gemma4VisionForward` puts every
+// one of them through `MakeDevBf16` before its first GEMM, so an f32 store would
+// cost 2x the checkpoint's bytes and buy nothing (#1359, and see
+// `qwen3_vl_vision.h` for the mirror argument and the bit-identity proof).
 struct Gemma4VisionBlockWeights {
-  std::vector<float> input_ln, post_attn_ln, pre_ff_ln, post_ff_ln;  // [H]
-  std::vector<float> q_proj, k_proj, v_proj, o_proj;  // [H,H]  (nh*head_dim==H)
-  std::vector<float> q_norm, k_norm;                  // [head_dim]
-  std::vector<float> gate_proj, up_proj;              // [I,H]
-  std::vector<float> down_proj;                       // [H,I]
+  std::vector<uint16_t> input_ln, post_attn_ln, pre_ff_ln, post_ff_ln;  // [H]
+  std::vector<uint16_t> q_proj, k_proj, v_proj, o_proj;  // [H,H]  (nh*head_dim==H)
+  std::vector<uint16_t> q_norm, k_norm;                  // [head_dim]
+  std::vector<uint16_t> gate_proj, up_proj;              // [I,H]
+  std::vector<uint16_t> down_proj;                       // [H,I]
   // QAT clamps (q/k/v share in-clamp; gate/up share in- AND out-clamp).
   Clip q_clip, k_clip, v_clip, o_clip, gate_clip, up_clip, down_clip;
 };
 
 struct Gemma4VisionWeights {
-  std::vector<float> input_proj;              // [H, patch_dim]  (768x768)
-  std::vector<float> position_embedding_table;  // [2*pos_embed_size*H]
+  std::vector<uint16_t> input_proj;              // [H, patch_dim]  (768x768)
+  // f32, DELIBERATELY, and the only such field here — the same exception
+  // `Qwen3VLVisionWeights::pos_embed_w` carries. `Gemma4VisionForward` SUMS the
+  // x and y rows of this table on the host (`gemma4_vision.cpp:199-210`) and
+  // narrows only the sum, so the stored values reach arithmetic and narrowing
+  // the store would move the result. Reconciling it onto the mirror is owed
+  // with the Qwen3-VL one (`.agents/specs/vision-tower-dtype-polarity.md` §4.3).
+  std::vector<float> position_embedding_table;   // [2*pos_embed_size*H]
   std::vector<Gemma4VisionBlockWeights> blocks;  // depth
-  std::vector<float> embed_projection;        // [text_hidden, H]  (2560x768)
+  std::vector<uint16_t> embed_projection;        // [text_hidden, H]  (2560x768)
 };
 
 // Per-stage captures for the G2-impl unit gates (all host f32, valid rows only).
